Refactor: move config and config errors into sub mod "config"
